2003-12-23  Christian Stimming  <stimming@tuhh.de>
	* src/engine/gnc-trace.c: Patch enabling compilation on Mac OS X
	by Rich Johnson <rich@dogstar-interactive.com>: Some platforms
	define timeval data members as "int".  Explicitly casting to "long
	int" for "%ld" directive avoids type mismatch.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@9739 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/gnc-trace.c
@@ -206,7 +206,8 @@ gnc_report_clock (int clockno, gncModuleType module, gncLogLevel log_level,
   if (!fout) gnc_log_init();
 
   fprintf (fout, "Clock %d Elapsed: %ld.%06lds %s: ",
-           clockno, now.tv_sec, now.tv_usec, gnc_log_prettify (function_name));
+           clockno, (long int) now.tv_sec, (long int) now.tv_usec, 
+	   gnc_log_prettify (function_name));
 
   va_start (ap, format);
 
@@ -237,8 +238,8 @@ gnc_report_clock_total (int clockno,
 
   fprintf (fout, "Clock %d Total Elapsed: %ld.%06lds  %s: ",
            clockno,
-           gnc_clock_total[clockno].tv_sec,
-           gnc_clock_total[clockno].tv_usec,
+           (long int) gnc_clock_total[clockno].tv_sec,
+           (long int) gnc_clock_total[clockno].tv_usec,
            gnc_log_prettify (function_name));
 
   va_start (ap, format);
